chore(jsnam) : 파일별 분류
