+++ src/entities/community/model/model.tsx
@@ -1,0 +1,168 @@
+import image from 'shared/assets/image/main/image1.png';
+
+import { CommunityItem } from './types';
+
+export const COMMUNITY_ITEM: CommunityItem = {
+  CommunityListInfo: [
+    {
+      id: '1',
+      songinfo: 'Pink! · 권진아',
+      title: '텍스트 초과 시 줄임표 뜨도록!!!!!!!!',
+      rateCount: '1',
+      reviewCount: '25',
+      img: image,
+      tag: [
+        {
+          id: 'one',
+          name: 'K-POP',
+          type: 'genre',
+        },
+        {
+          id: 'two',
+          name: '귀여운',
+          type: 'mood',
+        },
+      ],
+    },
+    {
+      id: '2',
+      songinfo: 'Butter · BTS',
+      title: '이 곡은 정말 대단해요!',
+      rateCount: '5',
+      reviewCount: '40',
+      img: image,
+      tag: [
+        {
+          id: 'one',
+          name: 'Pop',
+          type: 'genre',
+        },
+        {
+          id: 'two',
+          name: '댄스',
+          type: 'mood',
+        },
+      ],
+    },
+    {
+      id: '3',
+      songinfo: 'LALISA · Lisa',
+      title: 'LALISA 노래 너무 좋다!',
+      rateCount: '4',
+      reviewCount: '30',
+      img: image,
+      tag: [
+        {
+          id: 'one',
+          name: '힙합',
+          type: 'genre',
+        },
+        {
+          id: 'two',
+          name: '강렬한',
+          type: 'mood',
+        },
+      ],
+    },
+    {
+      id: '4',
+      songinfo: 'Love Dive · IVE',
+      title: '몽환적인 느낌의 곡입니다.',
+      rateCount: '4',
+      reviewCount: '20',
+      img: image,
+      tag: [
+        {
+          id: 'one',
+          name: '발라드',
+          type: 'genre',
+        },
+        {
+          id: 'two',
+          name: '몽환적인',
+          type: 'mood',
+        },
+      ],
+    },
+    {
+      id: '5',
+      songinfo: 'Hype Boy · NewJeans',
+      title: '이번 곡 대세네요!',
+      rateCount: '5',
+      reviewCount: '50',
+      img: image,
+      tag: [
+        {
+          id: 'one',
+          name: '신나는',
+          type: 'mood',
+        },
+        {
+          id: 'two',
+          name: '트렌디',
+          type: 'mood',
+        },
+      ],
+    },
+    {
+      id: '6',
+      songinfo: 'Next Level · aespa',
+      title: '이 곡으로 세계관이 확장됩니다.',
+      rateCount: '4',
+      reviewCount: '35',
+      img: image,
+      tag: [
+        {
+          id: 'one',
+          name: '강렬한',
+          type: 'mood',
+        },
+        {
+          id: 'two',
+          name: '미래적',
+          type: 'mood',
+        },
+      ],
+    },
+    {
+      id: '7',
+      songinfo: 'Ditto · NewJeans',
+      title: '잔잔하면서도 중독성이 있어요.',
+      rateCount: '4',
+      reviewCount: '18',
+      img: image,
+      tag: [
+        {
+          id: 'one',
+          name: '잔잔한',
+          type: 'mood',
+        },
+        {
+          id: 'two',
+          name: '트렌디',
+          type: 'mood',
+        },
+      ],
+    },
+    {
+      id: '8',
+      songinfo: 'Attention · NewJeans',
+      title: '들으면 집중되는 매력이 있어요.',
+      rateCount: '4',
+      reviewCount: '45',
+      img: image,
+      tag: [
+        {
+          id: 'one',
+          name: '중독성',
+          type: 'mood',
+        },
+        {
+          id: 'two',
+          name: '트렌디',
+          type: 'mood',
+        },
+      ],
+    },
+  ],
+};
