+++ src/entities/community/model/types.tsx
