+++ src/entities/community/ui/RecommendedPost.tsx
