+++ src/entities/home/model/model.ts
@@ -1,0 +1,217 @@
+import image from 'shared/assets/image/main/image1.png';
+import image2 from 'shared/assets/image/main/roundimg.png';
+
+import { MainItem } from './types';
+
+export const MAIN_ITEM: MainItem = {
+  GenreMusicList: [
+    {
+      id: '1',
+      title: 'Honey',
+      name: 'TRPP',
+      img: image,
+    },
+    {
+      id: '2',
+      title: 'Honey',
+      name: 'TRPP',
+      img: image,
+    },
+    {
+      id: '3',
+      title: 'Honey',
+      name: 'TRPP',
+      img: image,
+    },
+    {
+      id: '4',
+      title: 'Honey',
+      name: 'TRPP',
+      img: image,
+    },
+  ],
+
+  LikeMusicList: [
+    {
+      id: '1',
+      title: 'Honey',
+      name: 'TRPP',
+      img: image,
+    },
+    {
+      id: '2',
+      title: 'Honey',
+      name: 'TRPP',
+      img: image,
+    },
+    {
+      id: '3',
+      title: 'Honey',
+      name: 'TRPP',
+      img: image,
+    },
+    {
+      id: '4',
+      title: 'Honey',
+      name: 'TRPP',
+      img: image,
+    },
+    {
+      id: '5',
+      title: 'Honey',
+      name: 'TRPP',
+      img: image,
+    },
+    {
+      id: '6',
+      title: 'Honey',
+      name: 'TRPP',
+      img: image,
+    },
+    {
+      id: '7',
+      title: 'Honey',
+      name: 'TRPP',
+      img: image,
+    },
+    {
+      id: '8',
+      title: 'Honey',
+      name: 'TRPP',
+      img: image,
+    },
+    {
+      id: '7',
+      title: 'Honey',
+      name: 'TRPP',
+      img: image,
+    },
+    {
+      id: '8',
+      title: 'Honey',
+      name: 'TRPP',
+      img: image,
+    },
+  ],
+
+  HotMusicList: [
+    {
+      id: '1',
+      title: '그린내',
+      name: '실리카겔',
+      img: image2,
+    },
+    {
+      id: '2',
+      title: '그린내',
+      name: '실리카겔',
+      img: image2,
+    },
+    {
+      id: '3',
+      title: '그린내',
+      name: '실리카겔',
+      img: image2,
+    },
+    {
+      id: '4',
+      title: '그린내',
+      name: '실리카겔',
+      img: image2,
+    },
+    {
+      id: '5',
+      title: '그린내',
+      name: '실리카겔',
+      img: image2,
+    },
+    {
+      id: '6',
+      title: '그린내',
+      name: '실리카겔',
+      img: image2,
+    },
+  ],
+
+  CommunityMusicInfo: {
+    id: '1',
+    img: image,
+    title: '눈 · 자이언티',
+    description: '첫눈이 오면 제일 먼저 듣고 싶은 노래',
+    communityList: [
+      {
+        id: '1',
+        img: image,
+        date: '2024-12-10 00:12',
+        title: 'Highway Tune · Greta Van Fleet',
+        description: 'max-width 396이고 초과시 ...으로 나오도록 설정',
+        comment: '7',
+        recommend: '1',
+        views: '14',
+        user: '남재상',
+      },
+      {
+        id: '2',
+        img: image,
+        date: '2024-12-09 17:00',
+        title: 'Highway Tune · Greta Van Fleet',
+        description: 'max-width 396이고 초과시 ...으로 나오도록 설정',
+        comment: '7',
+        recommend: '1',
+        views: '14',
+        user: '여기도 max-width 200으로 초과시',
+      },
+      {
+        id: '3',
+        img: image,
+        date: '2024-12-08 17:00',
+        title: 'Highway Tune · Greta Van Fleet',
+        description: 'max-width 396이고 초과시 ...으로 나오도록 설정',
+        comment: '7',
+        recommend: '1',
+        views: '14',
+        user: '여기도 max-width 200으로 초과시',
+      },
+      {
+        id: '4',
+        img: image,
+        date: '2024-12-01 17:00',
+        title: 'Highway Tune · Greta Van Fleet',
+        description: 'max-width 396이고 초과시 ...으로 나오도록 설정',
+        comment: '7',
+        recommend: '1',
+        views: '14',
+        user: '여기도 max-width 200으로 초과시',
+      },
+      {
+        id: '5',
+        img: image,
+        date: '2023-12-06 17:00',
+        title: 'Highway Tune · Greta Van Fleet',
+        description: 'max-width 396이고 초과시 ...으로 나오도록 설정',
+        comment: '7',
+        recommend: '1',
+        views: '14',
+        user: '여기도 max-width 200으로 초과시',
+      },
+    ],
+  },
+
+  RecommendedMusicList: [
+    {
+      id: '1',
+      name: 'hathaw9y',
+      img: image,
+    },
+    {
+      id: '2',
+      name: 'Midnight jogging c',
+      img: image,
+    },
+    {
+      id: '3',
+      name: '모스크바서핑클럽',
+      img: image,
+    },
+  ],
+};
