+++ src/entities/home/model/types.ts
