+++ src/entities/memberInfo/review/ui/index.tsx
@@ -1,0 +1,18 @@
+import styled from '@emotion/styled';
+
+import { MAIN_ITEM } from 'entities/home/model/model';
+
+import { MemberReview } from './MemberReview';
+
+export const MemberInfoReview = () => {
+  const MainContents = styled.div`
+    width: 100%;
+    height: 2832px;
+  `;
+
+  return (
+    <MainContents>
+      <MemberReview CommunityMusicInfo={MAIN_ITEM.CommunityMusicInfo}></MemberReview>
+    </MainContents>
+  );
+};
