+++ src/pages/community/index.tsx
@@ -1,6 +1,6 @@
-import { Community } from 'pages/community/ui';
-
 import { MainLayout } from 'widgets/ui/';
+
+import { Community } from 'entities/community/ui/';
 
 const CommunityPage = () => {
   return (
