+++ src/pages/demo/index.tsx
@@ -1,10 +1,10 @@
 import { useState } from 'react';
 
-import { Main } from 'pages/home/ui';
-
 import { MainLayout } from 'widgets/ui/';
 
 import { MusicSelectionModal } from 'features/musicPreference/ui/MusicPreferenceModal';
+
+import { Main } from 'entities/home/ui';
 
 // 컴포넌트 렌더링 테스트를 위한 /demo 페이지
 const Demo = () => {
