+++ src/pages/home/index.tsx
@@ -4,7 +4,7 @@ import { MainLayout } from 'widgets/ui/';
 
 // import { MusicSelectionModal } from 'features/musicPreference/ui/MusicPreferenceModal';
 
-import { Main } from './ui';
+import { Main } from 'entities/home/ui/index';
 
 // 컴포넌트 렌더링 테스트를 위한 /demo 페이지
 const Home = () => {
