+++ src/pages/memberInfo/community.tsx
@@ -1,0 +1,11 @@
+import { MainLayout } from 'widgets/ui/Layout';
+
+import { MemberInfoCommunity } from 'entities/memberInfo/community/ui';
+
+export const MemberCommunity = () => {
+  return (
+    <MainLayout>
+      <MemberInfoCommunity></MemberInfoCommunity>
+    </MainLayout>
+  );
+};
