--- src/pages/memberInfo/community/index.tsx
@@ -1,11 +1,0 @@
-import { MainLayout } from 'widgets/ui/Layout';
-
-import { MemberInfoCommunity } from './ui';
-
-export const MemberCommunity = () => {
-  return (
-    <MainLayout>
-      <MemberInfoCommunity></MemberInfoCommunity>
-    </MainLayout>
-  );
-};
