--- src/pages/memberInfo/community/ui/index.tsx
@@ -1,18 +1,0 @@
-import styled from '@emotion/styled';
-
-import { MAIN_ITEM } from 'pages/home/ui/model';
-
-import { MemberCommunity } from './MemberCommunity';
-
-export const MemberInfoCommunity = () => {
-  const MainContents = styled.div`
-    width: 100%;
-    height: 2832px;
-  `;
-
-  return (
-    <MainContents>
-      <MemberCommunity CommunityMusicInfo={MAIN_ITEM.CommunityMusicInfo}></MemberCommunity>
-    </MainContents>
-  );
-};
