+++ src/pages/memberInfo/index.tsx
@@ -1,3 +1,3 @@
-export { MemberCommunity } from './community/index';
-export { MemberInfo } from './member/index';
-export { MemberReview } from './review/index';
+export { MemberCommunity } from './community';
+export { MemberInfo } from './member';
+export { MemberReview } from './review';
