+++ src/pages/memberInfo/member.tsx
@@ -1,0 +1,11 @@
+import { MainLayout } from 'widgets/ui/Layout';
+
+import { MemberInfoPage } from 'features/memberInfo/member/ui';
+
+export const MemberInfo = () => {
+  return (
+    <MainLayout>
+      <MemberInfoPage></MemberInfoPage>
+    </MainLayout>
+  );
+};
