+++ src/pages/memberInfo/review.tsx
@@ -1,0 +1,11 @@
+import { MainLayout } from 'widgets/ui/Layout';
+
+import { MemberInfoReview } from 'entities/memberInfo/review/ui';
+
+export const MemberReview = () => {
+  return (
+    <MainLayout>
+      <MemberInfoReview></MemberInfoReview>
+    </MainLayout>
+  );
+};
