--- src/pages/memberInfo/review/index.tsx
@@ -1,11 +1,0 @@
-import { MainLayout } from 'widgets/ui/Layout';
-
-import { MemberInfoReview } from './ui';
-
-export const MemberReview = () => {
-  return (
-    <MainLayout>
-      <MemberInfoReview></MemberInfoReview>
-    </MainLayout>
-  );
-};
